change db + adding test

+++ src/main/java/com/service/inspection/repositories/EquipmentRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EquipmentRepository extends JpaRepository<Equipment, Integer> {
+public interface EquipmentRepository extends JpaRepository<Equipment, Long> {
 
 }
